Add drawing_area parameter to API
Add drawing_area parameter in activateSurface API
and syncDraw event.
This is needed for the following feature
 - To request writting area
 - To inform application of writting area

I assume that
"normal.full" is a full size area in application area,
and "split.[main|sub]" is an upper/lower half size area
in application area.

For example, when Navigation app receives syncDraw event
which the label is Navigation, and the area name is
"normal.full", it means Navigation will be displayed
as a full size(in application area).

When Navigation app receives syncDraw event which
the label is same, and the area name is "split.main",
it means Navigation will be displayed
as an upper half size(in application area).
Application must return endDraw after changing its design
according to the size,
when the application receives syncDraw event.

Layout is defined in layers.json file.
Current layout is set to full screen.
Bug-AGL: SPEC-987

Change-Id: I5efed3dda88275e18dc1d951b98bdfcfc79ee46b
Signed-off-by: Kazumasa Mitsunari <knimitz@witz-inc.co.jp>