Add background layer for CES DEMO
MediaPlayer and Radio can not sound
when their surface have visibility=0.

And Navigation can not be displayed correctly
when the application is switched from visibility=0 to 1.

Because when visibility=0, weston event can not be catched.
So when the applications is not displayed,
they have visibility=1 and are moved to background layer.

Change-Id: Idf236e703a1e74cf564a1b213e6e15fcfa42a29e
Signed-off-by: Yuta Doi <yuta-d@witz-inc.co.jp>