Bug Fix: Window Manager doesn't react after killing app process
After killing an application which has surfaces,
Window Manager releases the information for app.
Then, invalid access happens.
As a result, Application seems to loose their ability to display.
This is a potencial bug caused by upgrades of wayland-ivi-extension,
so this problem didn't occur in eel branch.
This patch fixes the bug.

Bug-AGL: SPEC-1421

Change-Id: If921dd4ea6bd46454356e97e1be7e5c136c7383b
Signed-off-by: Kazumasa Mitsunari <knimitz@witz-inc.co.jp>