Add signal when the displayed application changed
Add new signal "screen_updated" to Window Manager.
This is emitted when application size or visibility is changed,
the application ids are notified to subscribers.

There are some applications such like HomeScreen,
which needs the visibility of other applications turns on,
so add new event.

The key is "ids" and data is application id which is written in
config.xml for each application.

This signal format is here:
ON-EVENT windowmanager/screen_updated:
{
  "event":"windowmanager\/screen_updated",
  "data":{
    "ids":[
      "navigation",
      "hvac"
      ]
  },
  "jtype":"afb-event"
}

Bug-AGL: SPEC-1532

Change-Id: I6c1f263be00d4b496e0e0f77dd88b6118489772f
Signed-off-by: Kazumasa Mitsunari <knimitz@witz-inc.co.jp>