Convert the roles from old one to new
Update the role name (drawing_name)
and WindowManager converts the role from old one to new.

We plan to delete the old roles by GG.
Therefore the applications can use old olds until GG,
After GG, old roles can not be used.
If the applications use it,
WindowManager regards the applications as the role "fallback".
Therefore the applications may not be displayed in the intended layout.
For the role "fallback", please refer to JIRA SPEC-1331.

Bug-AGL: SPEC-1566

Change-Id: I6ca703183e3cab72ac460094ef14dc3675a98cb6
Signed-off-by: Yuta Doi <yuta-d@witz-inc.co.jp>