Rename APIs
rename WindowManager APIs as follows:
 - activateSurface   -> activateWindow
 - deactivateSurface -> deactivateWindow

The "surface" means information of display material frame buffer
managed by Graphics Subsystem (Weston).
And the "window" means the right to draw on display.

The "surface" is physical information and used inside WindowManager.
So we think it is not good for "surface" to be included in API name.
These APIs provide the function
which is to activate/deactivate the right to draw for the applications.
So we change to "activateWindow/deactivateWindow".

We plan to delete the old API by GG.
Therefore the applications can use old APIs until GG,
but the warning "-Wdeprecated-declarations" is occured when compile.
After GG, old APIs can not be used.
If use it, the error is occured when compile.

Bug-AGL: SPEC-1565

Change-Id: Ic7da4e5ddb759e642a2b60ecf61e327aaf74a699
Signed-off-by: Yuta Doi <yuta-d@witz-inc.co.jp>